费太高了

+++ src/main/java/com/bl/BL.java
@@ -19,7 +19,7 @@ import static com.bl.entity.ModEntities.EXPANDING_SPHERE;
 
 public class BL implements ModInitializer {
 	public static final String MOD_ID = "bl";
-	 public static void createExpandingSphere(float max, Vec3d pos, ServerWorld world) {
+	 public static void createExpandingSphere(double max, Vec3d pos, ServerWorld world) {
 		// 假设ExpandingSphereEntity已经定义
 		ExpandingSphereEntity sphere = new ExpandingSphereEntity(EXPANDING_SPHERE, world);
 		sphere.SetMax(max);
